FloatingPoint: Made float32 the extended format for float8.
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/fix_floatingpoint_calculator@46236 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/floatingpoint/MyFloat.java
@@ -83,7 +83,7 @@ public class MyFloat {
     public Format extended() {
       // TODO: Add support for arbitrary sizes
       if (equals(Format.Float8)) {
-        return Float16;
+        return Float32;
       }
       if (equals(Format.Float16)) {
         return Float32;
